+++ game/paddle.go
@@ -4,15 +4,13 @@ import "github.com/veandco/go-sdl2/sdl"
 
 type Paddle struct {
 	PosX, PosY    float64
-	Width, Height int
+	Width, Height float64
 }
 
-func (p *Paddle) Init(windowWidth, windowHeight int32) {
-	p.Width = int(windowWidth) / 11
-	p.Height = int(windowHeight) / 24
+func (p *Paddle) Init(windowWidth, windowHeight float64) {
+	p.Width = windowWidth / 11
+	p.Height = windowHeight / 24
 	p.PosY = float64(windowHeight - (windowHeight / 8))
-
-	//p.PosX = (int(windowWidth) / 2) - p.Width/2
 }
 
 func (p Paddle) Draw(renderer *sdl.Renderer) {
